+++ lib/ex_unit/test/ex_unit/callbacks_test.exs
@@ -3,33 +3,41 @@ Code.require_file "../../test_helper.exs", __FILE__
 defmodule ExUnit.CallbacksTest do
   use ExUnit.Case, sync: false
 
-  def setup_all do
+  setup_all do
     [my_context: :setup_all]
   end
 
-  def setup([my_context: context], test) do
+  setup do
+    [first_setup: true]
+  end
+
+  setup context do
     if Process.get(:ex_unit_callback) do
       raise "ex_unit_callback was not cleaned"
     else
-      Process.put(:ex_unit_callback, test)
+      Process.put(:ex_unit_callback, context[:test])
     end
 
-    assert context == :setup_all
+    assert context[:first_setup]
+    assert context[:my_context] == :setup_all
 
     [my_context: :setup]
   end
 
-  def teardown([my_context: context], test) do
-    assert context == :setup
-    assert Process.get(:ex_unit_callback) == test
+  teardown context do
+    assert context[:my_context] == :setup
+    assert Process.get(:ex_unit_callback) == context[:test]
     Process.delete(:ex_unit_callback)
+    []
   end
 
-  def teardown_all([my_context: context]) do
-    assert context == :setup_all
+  teardown_all context do
+    assert context[:my_context] == :setup_all
+    []
   end
 
-  test :callback do
+  test :callback, context do
+    assert context[:my_context] == :setup
     assert Process.get(:ex_unit_callback) == :test_callback
   end