+++ lib/mix/test/test_helper.exs
@@ -52,16 +52,15 @@ defmodule MixTest.Case do
       use ExUnit.Case, unquote(opts)
       import MixTest.Case
 
-      def teardown(_) do
+      teardown do
         Mix.env(:dev)
         Mix.Task.clear
         Mix.Shell.Process.flush
         Mix.Deps.Converger.clear_cache
         System.put_env("MIX_HOME", tmp_path(".mix"))
         del_tmp_paths
+        []
       end
-
-      defoverridable [teardown: 1]
     end
   end